refactor service create magic link
